finished the refactor of playlistDropdown
